Added a sendToTU() and deleteAllLists() function in piGetter
In rtuSender

I had to add a "masterList" as well as the indivudal sites lists. These are all list<String[]>. The reason for this is that because of how piGetter does its job, it fetches all the PI tags between two time intervals at once, this makes things difficult on how to handle its data. You cannot send them all at once to an RTu or else what will happen is that It will push out data to the sites in a syncrhnous order. You also can't read them at a common index because the data from PI is inconsistent in size between sites (ie KL1 returns 150 data points, but PS1 returns 90 data points).

I also added a removeList function and a bildMasterList() function as well as renaiming an old unused function to compoundJSON()