renamed the namespace
Created a csvOuptut class and have it output a csv document that shows the time stamps and potential values (kW).

also implemented a clear function in piGetter. Thsi will clear the valueList upon call. This is needed so you don't constantly add to the csv file upon restarting the time.